requirements.txt
streamlit
openai
bs4
faiss-cpu
python-dotenv
watchdog
streamlit
langchain_openai
langchain_core
python-dotenv
langchain_community
langserve
fastapi
uvicorn
bs4
pypdf
chromadb
faiss-cpu
langchainhub
groq
langchain-groq
sentence_transformers
langchain
huggingface_hub